+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.4.1] - 2020-01-11
 ### Fixed
+- Conflicts between files with same name and different extension
 
 ## [1.4.0] - 2020-01-06
 ### Added